requirements.txt
paho-mqtt
certifi
mysql-connector-python
sqlalchemy
python-dotenv
var_dump